Move sort into storage service, hook up manager and vdev tables, test
